remove all example files
