--- login-workflow/example/src/components/ForgotPasswordScreenBaseExample.tsx
@@ -1,6 +1,0 @@
-import React from 'react';
-import { ForgotPasswordScreen } from '@brightlayer-ui/react-native-auth-workflow';
-
-export const ForgotPasswordScreenBaseExample: React.FC = () => {
-    return <ForgotPasswordScreen />;
-};
