--- login-workflow/example/src/components/ResetPasswordScreenBaseExample.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-import { ResetPasswordScreen } from '@brightlayer-ui/react-native-auth-workflow';
-
-export const ResetPasswordScreenBaseExample: React.FC = () => (
-    <ResetPasswordScreen
-        accountParams={{
-            code: '123',
-            email: 'test@eaton.com',
-        }}
-    />
-);
