--- login-workflow/example/src/screens/index.tsx
@@ -1,2 +1,0 @@
-export * from './Home';
-export * from './Locations';
